requirements.txt
aiohttp
aiohttp-swagger
sqlalchemy
aio_manager
pytest-aiohttp
uvloop
alembic
asyncpg
asyncpgsa
marshmallow
webargs
psycopg2
pyjwt
aio-pika